coding for stream and program

+++ container/mpegts/src/pat.rs
@@ -14,7 +14,7 @@ pub struct Pat {
     version_number: u8,     //5bits
     continuity_counter: u8, //s4 bits
 
-    pub pmt_count: u8,
+    pub pmt_count: usize,
     pub pmt: [pmt::Pmt; 4],
 }
